refactored error class; modified it incorrectly earlier
