+++ src/aws_lambda_python/common_domain/validation_error.py
@@ -4,9 +4,3 @@ from pydantic import BaseModel
 class ValidationError(BaseModel):
     error_type: str | None = None
     error_message: str | None = None
-
-
-class CoordinationError(BaseModel):
-    error_type: str | None = None
-    error_message: str | None = None
-
