+++ src/aws_lambda_python/mpic_coordinator/mpic_coordinator.py
@@ -13,7 +13,7 @@ from aws_lambda_python.common_domain.check_response import CheckResponse, Annota
     CaaCheckResponseDetails, DcvCheckResponse, DcvCheckResponseDetails
 from aws_lambda_python.common_domain.check_request import CaaCheckRequest
 from aws_lambda_python.common_domain.check_request import DcvCheckRequest
-from aws_lambda_python.common_domain.errors import ValidationError
+from aws_lambda_python.common_domain.validation_error import ValidationError
 from aws_lambda_python.common_domain.enum.check_type import CheckType
 from aws_lambda_python.common_domain.messages.ErrorMessages import ErrorMessages
 from aws_lambda_python.mpic_coordinator.domain.mpic_request import MpicCaaRequest, MpicRequest, AnnotatedMpicRequest
